fix: wire non-blocking triage diff cap
