+++ test/unit/execution/build-plan-for-strategy.test.ts
@@ -11,14 +11,7 @@ import { afterEach, beforeEach, describe, expect, mock, test } from "bun:test";
 import type { TestStrategy } from "@/config/schema-types";
 import { ExecutionPlan, buildPlanForStrategy } from "@/execution";
 import type { PlanInputs } from "@/execution";
-import { _storyOrchestratorDeps } from "@/execution";
-import {
-  makeAutofixTestWriterStrategy,
-  makeDeclarationSink,
-} from "@/operations";
 import type { UserStory } from "@/prd/types";
-import type { NaxRuntime } from "@/runtime";
-import { _rollbackDeps } from "@/tdd";
 import { makeMockCallContext, makeMockPlanInputs, makeNaxConfig, makeStory, makeTestRuntime } from "@test/helpers";
 
 // ─────────────────────────────────────────────────────────────────────────────
@@ -478,759 +471,6 @@ describe("buildPlanForStrategy — AC3: new check phase wiring (US-005)", () => 
       "semantic-review",
       "adversarial-review",
     ]);
-  });
-});
-
-// ─────────────────────────────────────────────────────────────────────────────
-// US-005 AC4: fix strategy assembly in buildPlanForStrategy
-// ─────────────────────────────────────────────────────────────────────────────
-
-describe("buildPlanForStrategy — AC4: fix strategy assembly (US-005)", () => {
-  let capturedStrategyNames: string[] = [];
-  let origRunFixCycle: typeof _storyOrchestratorDeps.runFixCycle;
-  let origCallOp: typeof _storyOrchestratorDeps.callOp;
-  let origCaptureGitRef: typeof _storyOrchestratorDeps.captureGitRef;
-  let runtime: NaxRuntime;
-
-  beforeEach(() => {
-    capturedStrategyNames = [];
-    origRunFixCycle = _storyOrchestratorDeps.runFixCycle;
-    origCallOp = _storyOrchestratorDeps.callOp;
-    origCaptureGitRef = _storyOrchestratorDeps.captureGitRef;
-
-    _storyOrchestratorDeps.captureGitRef = mock(async () => "HEAD");
-    // Mock callOp: verifier returns failing output with test-runner finding so rectification triggers
-    _storyOrchestratorDeps.callOp = mock(async (_ctx: unknown, op: { name: string }) => {
-      if (op.name === "verifier") {
-        return { success: false, findings: [{ source: "test-runner", severity: "error", message: "test failed" }] };
-      }
-      return { success: true };
-    }) as typeof _storyOrchestratorDeps.callOp;
-    _storyOrchestratorDeps.runFixCycle = mock(async (cycle: { strategies: Array<{ name: string }> }) => {
-      capturedStrategyNames = cycle.strategies.map((s) => s.name);
-      return { iterations: [], finalFindings: [], exitReason: "no-strategy" as const, costUsd: 0 };
-    }) as typeof _storyOrchestratorDeps.runFixCycle;
-  });
-
-  afterEach(async () => {
-    _storyOrchestratorDeps.runFixCycle = origRunFixCycle;
-    _storyOrchestratorDeps.callOp = origCallOp;
-    _storyOrchestratorDeps.captureGitRef = origCaptureGitRef;
-    await runtime?.close();
-  });
-
-  function makeCtxWithRuntime(config = makeNaxConfig()) {
-    runtime = makeTestRuntime({ config });
-    return makeMockCallContext({ runtime });
-  }
-
-  test("AC4: lintFix command configured → mechanical-lintfix strategy assembled in rectification", async () => {
-    const story = makeStory({ attempts: 1 });
-    const config = makeNaxConfig({
-      quality: { commands: { lintFix: "bun run lint:fix" } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    expect(capturedStrategyNames).toContain("mechanical-lintfix");
-  });
-
-  test("AC4: formatFix command configured → mechanical-formatfix strategy assembled in rectification", async () => {
-    const story = makeStory({ attempts: 1 });
-    const config = makeNaxConfig({
-      quality: { commands: { formatFix: "bun run format:fix" } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    expect(capturedStrategyNames).toContain("mechanical-formatfix");
-  });
-
-  test("AC4: non-TDD verify-scoped failure still enters rectification", async () => {
-    _storyOrchestratorDeps.callOp = mock(async (_ctx: unknown, op: { name: string }) => {
-      if (op.name === "verify-scoped") {
-        return {
-          success: false,
-          findings: [{ source: "test-runner", severity: "error", message: "scoped test failed" }],
-        };
-      }
-      return { success: true };
-    }) as typeof _storyOrchestratorDeps.callOp;
-
-    const story = makeStory();
-    const config = makeNaxConfig({
-      quality: { commands: { lintFix: "bun run lint:fix" } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeNonTddInputs(story, {
-      verifyScoped: { workdir: "/tmp/test", storyId: story.id },
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "no-test", inputs);
-    await plan.run();
-    expect(capturedStrategyNames.length).toBeGreaterThan(0);
-  });
-
-  test("AC4: autofix enabled → autofix-implementer strategy assembled in rectification", async () => {
-    const story = makeStory({ attempts: 1 });
-    const config = makeNaxConfig({
-      quality: { autofix: { enabled: true } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    expect(capturedStrategyNames).toContain("autofix-implementer");
-  });
-
-  test("AC4: autofix enabled → autofix-test-writer strategy assembled in rectification", async () => {
-    const story = makeStory({ attempts: 1 });
-    const config = makeNaxConfig({
-      quality: { autofix: { enabled: true } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    expect(capturedStrategyNames).toContain("autofix-test-writer");
-  });
-
-  test("AC4: no fix commands + autofix disabled → no mechanical or autofix strategies assembled", async () => {
-    const story = makeStory({ attempts: 1 });
-    const config = makeNaxConfig({
-      quality: { commands: {}, autofix: { enabled: false } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    // Only the fullSuiteRectifyStrategy should be present (prepended by buildPlanForStrategy for TDD+gate)
-    expect(capturedStrategyNames).not.toContain("mechanical-lintfix");
-    expect(capturedStrategyNames).not.toContain("mechanical-formatfix");
-    expect(capturedStrategyNames).not.toContain("autofix-implementer");
-    expect(capturedStrategyNames).not.toContain("autofix-test-writer");
-  });
-
-  // Regression: single-session (tdd-simple / test-after / no-test) scoped test
-  // failures emit `source: "test-runner"` findings. full-suite-rectify is the only
-  // strategy whose appliesTo matches that source. Before the fix it was gated behind
-  // TDD / per-story regression, so a single-session scoped failure had NO matching
-  // strategy and the cycle exited "no-strategy" at iteration 0 — the story failed
-  // without one fix attempt. With autofix disabled and no fix commands, the ONLY
-  // strategy that should remain is full-suite-rectify.
-  test("regression: single-session + verifyScoped phase → full-suite-rectify assembled (was no-strategy)", async () => {
-    _storyOrchestratorDeps.callOp = mock(async (_ctx: unknown, op: { name: string }) => {
-      if (op.name === "verify-scoped") {
-        return {
-          success: false,
-          findings: [
-            { source: "test-runner", severity: "error", category: "failed-test", message: "scoped test failed" },
-          ],
-        };
-      }
-      return { success: true };
-    }) as typeof _storyOrchestratorDeps.callOp;
-
-    const story = makeStory();
-    const config = makeNaxConfig({
-      quality: { commands: {}, autofix: { enabled: false } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeNonTddInputs(story, {
-      verifyScoped: { workdir: "/tmp/test", storyId: story.id },
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "no-test", inputs);
-    await plan.run();
-    // autofix disabled + no fix commands → full-suite-rectify is the ONLY
-    // assemblable strategy. Exact-set assertion guards against an accidental
-    // extra strategy slipping in.
-    expect(capturedStrategyNames).toEqual(["full-suite-rectify"]);
-  });
-
-  // Negative guard: a single-session plan with NO verify-scoped phase and deferred
-  // regression must NOT load full-suite-rectify — there is no phase that can emit
-  // test-runner findings, so the strategy would be dead weight. The cycle is driven
-  // by a lint-check failure (mechanical-lintfix matches) to ensure runFixCycle runs.
-  test("single-session without verifyScoped phase → full-suite-rectify NOT assembled", async () => {
-    _storyOrchestratorDeps.callOp = mock(async (_ctx: unknown, op: { name: string }) => {
-      if (op.name === "lint-check") {
-        return { success: false, findings: [{ source: "lint", severity: "error", message: "lint failed" }] };
-      }
-      return { success: true };
-    }) as typeof _storyOrchestratorDeps.callOp;
-
-    const story = makeStory();
-    const config = makeNaxConfig({
-      quality: { commands: { lintFix: "bun run lint:fix" }, autofix: { enabled: false } },
-      execution: { regressionGate: { mode: "deferred" }, rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeNonTddInputs(story, {
-      lintCheck: { workdir: "/tmp/test", storyId: story.id },
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "no-test", inputs);
-    await plan.run();
-    expect(capturedStrategyNames).toContain("mechanical-lintfix");
-    expect(capturedStrategyNames).not.toContain("full-suite-rectify");
-  });
-
-  // No-regression guard: three-session TDD must STILL load full-suite-rectify
-  // exactly as before (the fix only widens single-session, never narrows TDD).
-  test("no-regression: three-session TDD still assembles full-suite-rectify", async () => {
-    const story = makeStory({ attempts: 1 });
-    const config = makeNaxConfig({
-      quality: { commands: {}, autofix: { enabled: false } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-    });
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    expect(capturedStrategyNames).toContain("full-suite-rectify");
-  });
-});
-
-// ─────────────────────────────────────────────────────────────────────────────
-// US-006 AC1: triage scope routes NBF strategies by fixTarget
-// ─────────────────────────────────────────────────────────────────────────────
-
-describe("buildPlanForStrategy — AC1: triage scope NBF strategy assembly (US-006)", () => {
-  let capturedStrategyNamesByCall: string[][] = [];
-  let origRunFixCycle: typeof _storyOrchestratorDeps.runFixCycle;
-  let origCallOp: typeof _storyOrchestratorDeps.callOp;
-  let origCaptureGitRef: typeof _storyOrchestratorDeps.captureGitRef;
-  let origRollbackSpawn: typeof _rollbackDeps.spawn;
-  let origRollbackAutoCommit: typeof _rollbackDeps.autoCommitIfDirty;
-  let runtime: NaxRuntime;
-
-  beforeEach(() => {
-    capturedStrategyNamesByCall = [];
-    origRunFixCycle = _storyOrchestratorDeps.runFixCycle;
-    origCallOp = _storyOrchestratorDeps.callOp;
-    origCaptureGitRef = _storyOrchestratorDeps.captureGitRef;
-    origRollbackSpawn = _rollbackDeps.spawn;
-    origRollbackAutoCommit = _rollbackDeps.autoCommitIfDirty;
-
-    _storyOrchestratorDeps.captureGitRef = mock(async () => "HEAD");
-    // captureSnapshotRef uses _rollbackDeps.spawn for git rev-parse HEAD.
-    _rollbackDeps.autoCommitIfDirty = mock(async () => {});
-    _rollbackDeps.spawn = mock((_cmd: string[], _opts: unknown) => ({
-      stdout: new Response("abc1234\n").body,
-      stderr: new Response("").body,
-      exited: Promise.resolve(0),
-    })) as typeof _rollbackDeps.spawn;
-    // Mock callOp: adversarial-review passes with advisory findings so NBF fires,
-    // verifier passes so adversarial-review actually runs (it sits AFTER verifier
-    // in canonical order), and full-suite-gate fails so the main rectification cycle
-    // has findings to consume.
-    _storyOrchestratorDeps.callOp = mock(async (_ctx: unknown, op: { name: string }) => {
-      if (op.name === "adversarial-review") {
-        return {
-          success: true,
-          passed: true,
-          advisoryFindings: [
-            {
-              source: "adversarial-review",
-              severity: "info",
-              category: "test-gap",
-              message: "advisory gap",
-              fixTarget: "test",
-            },
-          ],
-        };
-      }
-      if (op.name === "full-suite-gate") {
-        return {
-          success: false,
-          findings: [{ source: "test-runner", severity: "error", message: "test failed" }],
-        };
-      }
-      return { success: true };
-    }) as typeof _storyOrchestratorDeps.callOp;
-    _storyOrchestratorDeps.runFixCycle = mock(async (cycle: { strategies: Array<{ name: string }> }) => {
-      capturedStrategyNamesByCall.push(cycle.strategies.map((s) => s.name));
-      return { iterations: [], finalFindings: [], exitReason: "no-strategy" as const, costUsd: 0 };
-    }) as typeof _storyOrchestratorDeps.runFixCycle;
-  });
-
-  afterEach(async () => {
-    _storyOrchestratorDeps.runFixCycle = origRunFixCycle;
-    _storyOrchestratorDeps.callOp = origCallOp;
-    _storyOrchestratorDeps.captureGitRef = origCaptureGitRef;
-    _rollbackDeps.spawn = origRollbackSpawn;
-    _rollbackDeps.autoCommitIfDirty = origRollbackAutoCommit;
-    await runtime?.close();
-  });
-
-  // The NBF call is the LAST runFixCycle invocation; any subsequent runFixCycle
-  // calls come from post-rectification-resume retries, not NBF. NBF strategies
-  // are distinguished by their fixTarget-routed composition
-  // (adversarialReviewByFixTarget / includeAdversarialReview), not by name
-  // overlap. For AC1 we only need to assert name membership on the LAST call
-  // (which is NBF when NBF fires).
-  function lastCaptured(): string[] {
-    return capturedStrategyNamesByCall[capturedStrategyNamesByCall.length - 1] ?? [];
-  }
-
-  function makeCtxWithRuntime(config = makeNaxConfig()) {
-    runtime = makeTestRuntime({ config });
-    return makeMockCallContext({ runtime });
-  }
-
-  function withTriageNbf(extra: Record<string, unknown> = {}): ReturnType<typeof makeNaxConfig> {
-    return makeNaxConfig({
-      quality: { commands: {}, autofix: { enabled: true } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-      review: {
-        adversarial: {
-          model: "balanced",
-          diffMode: "ref",
-          rules: [],
-          timeoutMs: 600_000,
-          parallel: false,
-          maxConcurrentSessions: 2,
-          nonBlockingFix: {
-            enabled: true,
-            scope: "triage",
-            regressionAttempts: 1,
-            verifierGuard: true,
-            ...extra,
-          },
-        },
-      },
-    });
-  }
-
-  function withNbfScope(scope: "source" | "both" | "triage"): ReturnType<typeof makeNaxConfig> {
-    return makeNaxConfig({
-      quality: { commands: {}, autofix: { enabled: true } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-      review: {
-        adversarial: {
-          model: "balanced",
-          diffMode: "ref",
-          rules: [],
-          timeoutMs: 600_000,
-          parallel: false,
-          maxConcurrentSessions: 2,
-          nonBlockingFix: { enabled: true, scope, regressionAttempts: 1, verifierGuard: true },
-        },
-      },
-    });
-  }
-
-  test("AC1: NBF scope=triage assembles autofix-implementer, autofix-test-writer, full-suite-rectify", async () => {
-    const story = makeStory({ attempts: 1 });
-    const config = withTriageNbf();
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      adversarialReview: {
-        story,
-        adversarialConfig: config.review.adversarial!,
-        mode: config.review.adversarial!.diffMode,
-      },
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    // NBF is the LAST runFixCycle call (after main rect + post-rectification-resume).
-    expect(capturedStrategyNamesByCall.length).toBeGreaterThanOrEqual(2);
-    const nbfNames = lastCaptured();
-    expect(nbfNames).toContain("autofix-implementer");
-    expect(nbfNames).toContain("autofix-test-writer");
-    expect(nbfNames).toContain("full-suite-rectify");
-  });
-
-  test("AC1: NBF scope=triage does NOT regress — scope=both still assembles the same three", async () => {
-    const story = makeStory({ attempts: 1 });
-    const config = withNbfScope("both");
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      adversarialReview: {
-        story,
-        adversarialConfig: config.review.adversarial!,
-        mode: config.review.adversarial!.diffMode,
-      },
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    expect(capturedStrategyNamesByCall.length).toBeGreaterThanOrEqual(2);
-    const nbfNames = lastCaptured();
-    expect(nbfNames).toContain("autofix-implementer");
-    expect(nbfNames).toContain("autofix-test-writer");
-    expect(nbfNames).toContain("full-suite-rectify");
-  });
-
-  test("AC1: NBF scope=source still assembles only autofix-implementer + full-suite-rectify", async () => {
-    // This test overrides the shared callOp mock so the main loop completes
-    // successfully (gate passes, verifier passes, adversarial-review produces advisory
-    // findings). That lets NBF fire, which is the call whose strategy composition we
-    // actually want to inspect. With full-suite-gate failing (the shared mock), the
-    // main loop short-circuits before adversarial-review runs, so NBF never fires.
-    _storyOrchestratorDeps.callOp = mock(async (_ctx: unknown, op: { name: string }) => {
-      if (op.name === "adversarial-review") {
-        return {
-          success: true,
-          passed: true,
-          advisoryFindings: [
-            {
-              source: "adversarial-review",
-              severity: "info",
-              category: "test-gap",
-              message: "advisory gap",
-              fixTarget: "test",
-            },
-          ],
-        };
-      }
-      return { success: true };
-    }) as typeof _storyOrchestratorDeps.callOp;
-
-    const story = makeStory({ attempts: 1 });
-    const config = withNbfScope("source");
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      adversarialReview: {
-        story,
-        adversarialConfig: config.review.adversarial!,
-        mode: config.review.adversarial!.diffMode,
-      },
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-    expect(capturedStrategyNamesByCall.length).toBeGreaterThanOrEqual(1);
-    // NBF is the only runFixCycle call (no gate failure → no main rectification).
-    // For scope:source, NBF assembles implementer + full-suite-rectify, never test-writer.
-    const nbfNames = capturedStrategyNamesByCall[0] ?? [];
-    expect(nbfNames).toContain("autofix-implementer");
-    expect(nbfNames).toContain("full-suite-rectify");
-    expect(nbfNames).not.toContain("autofix-test-writer");
-  });
-});
-
-// ─────────────────────────────────────────────────────────────────────────────
-// US-006 AC2/AC3/AC4: triage strategy set routes findings by fixTarget
-// ─────────────────────────────────────────────────────────────────────────────
-//
-// The spec says "Given the `triage` strategy set... when each strategy's
-// `appliesTo` is evaluated, then..." — the set is the one built by
-// `buildPlanForStrategy` for `scope: "triage"`, not a manual mirror of the
-// option shapes. We drive through `buildPlanForStrategy` and capture the NBF
-// strategies from the `runFixCycle` mock.
-//
-// Mock setup contract: when full-suite-gate fails, the main loop short-circuits
-// before adversarial-review runs, so NBF never fires (advisory findings empty).
-// To force NBF to fire we override the shared callOp so the gate passes and
-// adversarial-review produces advisory findings — NBF then becomes the ONLY
-// runFixCycle call, and we can assert on its strategies directly.
-
-describe("buildPlanForStrategy — AC2/AC3/AC4: triage strategy predicate behavior (driven through plan)", () => {
-  let capturedStrategiesByCall: Array<Array<import("@/findings").FixStrategy<import("@/findings").Finding, any, any, any>>> = [];
-  let origRunFixCycle: typeof _storyOrchestratorDeps.runFixCycle;
-  let origCallOp: typeof _storyOrchestratorDeps.callOp;
-  let origCaptureGitRef: typeof _storyOrchestratorDeps.captureGitRef;
-  let origRollbackSpawn: typeof _rollbackDeps.spawn;
-  let origRollbackAutoCommit: typeof _rollbackDeps.autoCommitIfDirty;
-  let runtime: NaxRuntime;
-
-  beforeEach(() => {
-    capturedStrategiesByCall = [];
-    origRunFixCycle = _storyOrchestratorDeps.runFixCycle;
-    origCallOp = _storyOrchestratorDeps.callOp;
-    origCaptureGitRef = _storyOrchestratorDeps.captureGitRef;
-    origRollbackSpawn = _rollbackDeps.spawn;
-    origRollbackAutoCommit = _rollbackDeps.autoCommitIfDirty;
-
-    _storyOrchestratorDeps.captureGitRef = mock(async () => "HEAD");
-    _rollbackDeps.autoCommitIfDirty = mock(async () => {});
-    _rollbackDeps.spawn = mock((_cmd: string[], _opts: unknown) => ({
-      stdout: new Response("abc1234\n").body,
-      stderr: new Response("").body,
-      exited: Promise.resolve(0),
-    })) as typeof _rollbackDeps.spawn;
-    // Gate PASSES (so the main loop reaches adversarial-review), adversarial-review
-    // produces advisory findings, NBF fires. With gate failing, the main loop
-    // short-circuits and NBF never fires (see AC1: scope=source test for the
-    // matching pattern).
-    _storyOrchestratorDeps.callOp = mock(async (_ctx: unknown, op: { name: string }) => {
-      if (op.name === "adversarial-review") {
-        return {
-          success: true,
-          passed: true,
-          advisoryFindings: [
-            {
-              source: "adversarial-review",
-              severity: "info",
-              category: "test-gap",
-              message: "advisory gap",
-              fixTarget: "test",
-            },
-          ],
-        };
-      }
-      return { success: true };
-    }) as typeof _storyOrchestratorDeps.callOp;
-    _storyOrchestratorDeps.runFixCycle = mock(
-      async (cycle: { strategies: Array<import("@/findings").FixStrategy<import("@/findings").Finding, any, any, any>> }) => {
-        capturedStrategiesByCall.push(cycle.strategies);
-        return { iterations: [], finalFindings: [], exitReason: "no-strategy" as const, costUsd: 0 };
-      },
-    ) as typeof _storyOrchestratorDeps.runFixCycle;
-  });
-
-  afterEach(async () => {
-    _storyOrchestratorDeps.runFixCycle = origRunFixCycle;
-    _storyOrchestratorDeps.callOp = origCallOp;
-    _storyOrchestratorDeps.captureGitRef = origCaptureGitRef;
-    _rollbackDeps.spawn = origRollbackSpawn;
-    _rollbackDeps.autoCommitIfDirty = origRollbackAutoCommit;
-    await runtime?.close();
-  });
-
-  function makeCtxWithRuntime(config = makeNaxConfig()) {
-    runtime = makeTestRuntime({ config });
-    return makeMockCallContext({ runtime });
-  }
-
-  function withTriageNbf(): ReturnType<typeof makeNaxConfig> {
-    return makeNaxConfig({
-      quality: { commands: {}, autofix: { enabled: true } },
-      execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-      review: {
-        adversarial: {
-          model: "balanced",
-          diffMode: "ref",
-          rules: [],
-          timeoutMs: 600_000,
-          parallel: false,
-          maxConcurrentSessions: 2,
-          nonBlockingFix: {
-            enabled: true,
-            scope: "triage",
-            regressionAttempts: 1,
-            verifierGuard: true,
-          },
-        },
-      },
-    });
-  }
-
-  /**
-   * Returns the captured NBF strategy set. With the gate-passing callOp mock,
-   * NBF is the ONLY runFixCycle call (no gate failure → no main rect).
-   */
-  function nbfStrategies(): Array<import("@/findings").FixStrategy<import("@/findings").Finding, any, any, any>> {
-    if (capturedStrategiesByCall.length === 0) {
-      throw new Error("NBF never fired — capturedStrategiesByCall is empty");
-    }
-    return capturedStrategiesByCall[capturedStrategiesByCall.length - 1] ?? [];
-  }
-
-  function findStrategy(
-    name: string,
-    set: Array<import("@/findings").FixStrategy<import("@/findings").Finding, any, any, any>>,
-  ): import("@/findings").FixStrategy<import("@/findings").Finding, any, any, any> {
-    const s = set.find((x) => x.name === name);
-    if (!s) {
-      throw new Error(
-        `strategy '${name}' not found in captured NBF set: ${set.map((x) => x.name).join(", ")}`,
-      );
-    }
-    return s;
-  }
-
-  async function buildAndRunTriagePlan() {
-    const story = makeStory({ attempts: 1 });
-    const config = withTriageNbf();
-    const ctx = makeCtxWithRuntime(config);
-    const inputs = makeTddRetryInputs(story, {
-      adversarialReview: {
-        story,
-        adversarialConfig: config.review.adversarial!,
-        mode: config.review.adversarial!.diffMode,
-      },
-      rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-    });
-    const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-    await plan.run();
-  }
-
-  function makeFinding(overrides: Record<string, unknown> = {}): import("@/findings").Finding {
-    return {
-      source: "lint",
-      severity: "error",
-      category: "lint-error",
-      message: "msg",
-      fixTarget: "source",
-      ...overrides,
-    };
-  }
-
-  test("AC2: triage scope → implementer.appliesTo=true and test-writer.appliesTo=false for adversarial source finding", async () => {
-    await buildAndRunTriagePlan();
-    const nbfSet = nbfStrategies();
-    const implementer = findStrategy("autofix-implementer", nbfSet);
-    const testWriter = findStrategy("autofix-test-writer", nbfSet);
-    const finding = makeFinding({ source: "adversarial-review", fixTarget: "source" });
-    expect(implementer.appliesTo(finding)).toBe(true);
-    expect(testWriter.appliesTo(finding)).toBe(false);
-  });
-
-  test("AC3: triage scope → test-writer.appliesTo=true and implementer.appliesTo=false for adversarial test finding", async () => {
-    await buildAndRunTriagePlan();
-    const nbfSet = nbfStrategies();
-    const implementer = findStrategy("autofix-implementer", nbfSet);
-    const testWriter = findStrategy("autofix-test-writer", nbfSet);
-    const finding = makeFinding({ source: "adversarial-review", fixTarget: "test" });
-    expect(testWriter.appliesTo(finding)).toBe(true);
-    expect(implementer.appliesTo(finding)).toBe(false);
-  });
-
-  test("AC4: triage scope → test-writer.appliesTo=true for advisory convention finding with fixTarget=test", async () => {
-    await buildAndRunTriagePlan();
-    const nbfSet = nbfStrategies();
-    const testWriter = findStrategy("autofix-test-writer", nbfSet);
-    const finding = makeFinding({
-      source: "adversarial-review",
-      category: "convention",
-      fixTarget: "test",
-    });
-    expect(testWriter.appliesTo(finding)).toBe(true);
-  });
-});
-
-// ─────────────────────────────────────────────────────────────────────────────
-// US-006 AC5/AC6: default-preserving factory options preserve existing behaviour
-// ─────────────────────────────────────────────────────────────────────────────
-//
-// AC5 is a factory-level test (per spec: "Given `makeAutofixTestWriterStrategy`
-// built with default options...") and stays as a direct factory test.
-//
-// AC6 is an integration test: the spec says "Given the blocking three-session
-// strategy set" — that's the set built by `buildPlanForStrategy` for three-session
-// TDD with NBF disabled (the "blocking" qualifier means before NBF, i.e. the
-// regular rectification cycle's strategies). We drive through `buildPlanForStrategy`
-// and capture the main-rectification strategies from the `runFixCycle` mock.
-// With the gate failing, main rect fires; with NBF disabled, NBF doesn't fire,
-// so the only captured call is the main rect.
-
-describe("buildPlanForStrategy — AC5/AC6: default-preserving factory options + blocking set behaviour", () => {
-  test("AC5: default makeAutofixTestWriterStrategy still claims adversarial source finding (preserves blanket behaviour)", () => {
-    const testWriter = makeAutofixTestWriterStrategy(makeStory(), makeNaxConfig(), makeDeclarationSink());
-    const finding: import("@/findings").Finding = {
-      source: "adversarial-review",
-      severity: "warning",
-      category: "input",
-      message: "advisory finding",
-      fixTarget: "source",
-    };
-    expect(testWriter.appliesTo(finding)).toBe(true);
-  });
-
-  test("AC6: blocking three-session set → test-writer.appliesTo=true and implementer.appliesTo=false for adversarial source finding (driven through plan)", async () => {
-    // Drive through buildPlanForStrategy with NBF disabled and gate failing,
-    // so the main-rectification cycle is the only runFixCycle call. The
-    // blocking three-session set assembles (per build-plan-for-strategy.ts:188-198):
-    //   - autofix-implementer with includeAdversarialReview: false (three-session)
-    //   - autofix-test-writer with default opts (blanket adversarial preserved)
-    //   - full-suite-rectify
-    const capturedStrategySets: Array<Array<import("@/findings").FixStrategy<import("@/findings").Finding, any, any, any>>> = [];
-    const origRunFixCycle = _storyOrchestratorDeps.runFixCycle;
-    const origCallOp = _storyOrchestratorDeps.callOp;
-    const origCaptureGitRef = _storyOrchestratorDeps.captureGitRef;
-    const origRollbackSpawn = _rollbackDeps.spawn;
-    const origRollbackAutoCommit = _rollbackDeps.autoCommitIfDirty;
-    let runtime: NaxRuntime;
-    try {
-      _storyOrchestratorDeps.captureGitRef = mock(async () => "HEAD");
-      _rollbackDeps.autoCommitIfDirty = mock(async () => {});
-      _rollbackDeps.spawn = mock((_cmd: string[], _opts: unknown) => ({
-        stdout: new Response("abc1234\n").body,
-        stderr: new Response("").body,
-        exited: Promise.resolve(0),
-      })) as typeof _rollbackDeps.spawn;
-      // Gate fails → main rect fires; NBF is disabled → only main rect runs.
-      _storyOrchestratorDeps.callOp = mock(async (_ctx: unknown, op: { name: string }) => {
-        if (op.name === "full-suite-gate") {
-          return {
-            success: false,
-            findings: [{ source: "test-runner", severity: "error", message: "test failed" }],
-          };
-        }
-        return { success: true };
-      }) as typeof _storyOrchestratorDeps.callOp;
-      _storyOrchestratorDeps.runFixCycle = mock(
-        async (cycle: { strategies: Array<import("@/findings").FixStrategy<import("@/findings").Finding, any, any, any>> }) => {
-          capturedStrategySets.push(cycle.strategies);
-          return { iterations: [], finalFindings: [], exitReason: "no-strategy" as const, costUsd: 0 };
-        },
-      ) as typeof _storyOrchestratorDeps.runFixCycle;
-
-      const config = makeNaxConfig({
-        quality: { commands: {}, autofix: { enabled: true } },
-        execution: { rectification: { enabled: true, maxAttemptsTotal: 2 } },
-        // No review.adversarial configured → NBF never enabled → blocking three-session set.
-      });
-      const story = makeStory({ attempts: 1 });
-      runtime = makeTestRuntime({ config });
-      const ctx = makeMockCallContext({ runtime });
-      const inputs = makeTddRetryInputs(story, {
-        rectification: { maxAttempts: 2, strategies: [], abortOnIncreasingFailures: false },
-      });
-      const plan = await buildPlanForStrategy(ctx, story, config, "three-session-tdd", inputs);
-      await plan.run();
-
-      // Main rect is the only runFixCycle call (gate fails → main rect; NBF disabled).
-      const mainRectSet = capturedStrategySets[capturedStrategySets.length - 1] ?? [];
-      expect(mainRectSet.length).toBeGreaterThan(0);
-      const implementer = mainRectSet.find((s) => s.name === "autofix-implementer");
-      const testWriter = mainRectSet.find((s) => s.name === "autofix-test-writer");
-      expect(implementer).toBeDefined();
-      expect(testWriter).toBeDefined();
-      const finding: import("@/findings").Finding = {
-        source: "adversarial-review",
-        severity: "warning",
-        category: "input",
-        message: "advisory finding",
-        fixTarget: "source",
-      };
-      // test-writer (default opts, blanket adversarial preserved) claims the finding.
-      expect(testWriter!.appliesTo(finding)).toBe(true);
-      // implementer (includeAdversarialReview: false for three-session) does NOT.
-      expect(implementer!.appliesTo(finding)).toBe(false);
-    } finally {
-      _storyOrchestratorDeps.runFixCycle = origRunFixCycle;
-      _storyOrchestratorDeps.callOp = origCallOp;
-      _storyOrchestratorDeps.captureGitRef = origCaptureGitRef;
-      _rollbackDeps.spawn = origRollbackSpawn;
-      _rollbackDeps.autoCommitIfDirty = origRollbackAutoCommit;
-      await runtime?.close();
-    }
   });
 });
